Geometry class (S4) and methods

A geom stores a table of points, a table of feature to which the points are associated and a table of groups, to which features are associated. A geom can be spatial (if it has a coordinate reference system assigned to it), but is not by default.

Details

A geom has one of three geometry types:

"point", when none of the points are connected to other points,

"line", where points with the same fid are connected following the sequence of their order, without the line closing in itself and

"polygon", where points with the same fid are connected following the sequence of their order and the line closes in on itself due to first and last point being the same. Moreover, polygon objects can contain holes.

The data model for storing points follows the spaghetti model. Points are stored as a sequence of x and y values, associated to a feature ID. The feature ID relates coordinates to features and thus common attributes. Points and Lines are implemented straightforward in this model, but polygons, which may contain holes, are a bit trickier. In geometr they are implemented as follows:

All points with the same fid make up one polygon, irrespective of it containing holes or not.

The outer path/ring of a polygon is composed of all points until a duplicated of its first point occurs. This signals that all following points are part of another path/ring, which must be inside the outer path and which consists of all points until a duplicate of it's first point occurs.

This repeats until all points of the feature are processed.

Moreover, a geom has a reference window, which is sort of a second extent that may be bigger (or smaller) than the extent and which determines the relative position of the points when plotting.

Slots

name

character(1)

classes

data.frame(.)

sources

data.frame(.)

concepts

data.frame(.)

labels

data.frame(.)

mappings

character(1)