Get a concept in an ontology

Get a concept in an ontology

get_concept(
  x = NULL,
  ...,
  regex = FALSE,
  tree = FALSE,
  na.rm = FALSE,
  mappings = "none",
  ontology = NULL
)

Arguments

x

character(1)
a table containing all columns of the ontology that shall be filter by the values in those columns.

...

combination of column name and value to filter that column by. The value to filter by can be provided as regular expression.

regex

logical(1)
whether or not the value in ... shall be matched in full, or whether any partial match should be returned.

tree

logical(1)
whether or not to output the whole ontology tree starting from the given search terms.

na.rm

logical(1)
whether or not missing concepts are omitted or not.

mappings

character(1)
the type of mappings within which to search for the concepts, possible values are "none", "all", "close", "broader", "narrower", "exact" or any combination thereof.

ontology

ontology(1)
either a path where the ontology is stored, or an already loaded ontology.

Value

A table of a subset of the ontology according to the values in ...

Examples

ontoDir <- system.file("extdata", "crops.rds", package = "ontologics")
onto <- load_ontology(path = ontoDir)

# exact matches from a loaded ontology ...
get_concept(x = data.frame(label = "FODDER CROPS"), ontology = onto)
#>          label class  id has_broader description
#> 1 FODDER CROPS group .05        <NA>        <NA>

# ... or one stored on the harddisc
get_concept(x = data.frame(label = "FODDER CROPS"), ontology = ontoDir)
#>          label class  id has_broader description
#> 1 FODDER CROPS group .05        <NA>        <NA>

# use regular expressions ...
get_concept(x = data.frame(label = "/*crops"), regex = TRUE, ontology = onto)
#> # A tibble: 12 × 9
#>    label             class id    has_b…¹ descr…² has_c…³ has_n…⁴ has_b…⁵ has_e…⁶
#>    <chr>             <chr> <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 Other bioenergy … class .01.… .01     NA      NA      NA      NA      NA     
#>  2 Fibre crops       class .03.… .03     NA      NA      NA      NA      NA     
#>  3 Woody flower cro… class .04.… .04     NA      NA      NA      NA      NA     
#>  4 Grass crops       class .05.… .05     NA      NA      NA      NA      NA     
#>  5 Other fodder cro… class .05.… .05     NA      NA      NA      NA      NA     
#>  6 Leguminous crops  class .08.… .08     NA      NA      NA      NA      NA     
#>  7 Medicinal crops   class .09.… .09     NA      NA      NA      NA      NA     
#>  8 Oilseed crops     class .11.… .11     NA      NA      NA      NA      NA     
#>  9 Other crops       class .12.… .12     NA      NA      NA      NA      NA     
#> 10 Stimulant crops   class .14.… .14     NA      NA      NA      NA      NA     
#> 11 Spice crops       class .14.… .14     NA      NA      NA      NA      NA     
#> 12 Other sugar crops class .15.… .15     NA      NA      NA      NA      NA     
#> # … with abbreviated variable names ¹​has_broader, ²​description,
#> #   ³​has_close_match, ⁴​has_narrower_match, ⁵​has_broader_match, ⁶​has_exact_match

get_concept(x = data.frame(label = "/*crops"), has_broader = ".05", regex = TRUE, ontology = onto)
#> # A tibble: 2 × 9
#>   label              class id    has_b…¹ descr…² has_c…³ has_n…⁴ has_b…⁵ has_e…⁶
#>   <chr>              <chr> <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 Grass crops        class .05.… .05     NA      NA      NA      NA      NA     
#> 2 Other fodder crops class .05.… .05     NA      NA      NA      NA      NA     
#> # … with abbreviated variable names ¹​has_broader, ²​description,
#> #   ³​has_close_match, ⁴​has_narrower_match, ⁵​has_broader_match, ⁶​has_exact_match

# get all concepts that are nested into another concept
get_concept(x = data.frame(label = "FODDER CROPS"), tree = TRUE, ontology = onto)
#> # A tibble: 4 × 9
#>   label              class id    has_b…¹ descr…² has_c…³ has_n…⁴ has_b…⁵ has_e…⁶
#>   <chr>              <chr> <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#> 1 FODDER CROPS       group .05   NA      NA      NA      NA      NA      NA     
#> 2 Grass crops        class .05.… .05     NA      NA      NA      NA      NA     
#> 3 Fodder legumes     class .05.… .05     NA      NA      NA      NA      NA     
#> 4 Other fodder crops class .05.… .05     NA      NA      NA      NA      NA     
#> # … with abbreviated variable names ¹​has_broader, ²​description,
#> #   ³​has_close_match, ⁴​has_narrower_match, ⁵​has_broader_match, ⁶​has_exact_match