Get a concept in an ontology

Get a concept in an ontology

get_concept(
  ...,
  regex = FALSE,
  external = FALSE,
  match.input = TRUE,
  ontology = NULL
)

Arguments

...

combination of column name and value to filter that column by. The value to filter by can be provided as regular expression, if regex = TRUE.

regex

logical(1)
if regex = TRUE, the columns defined in table are filtered by str_detect on the column values (if you define several, they are combined with an AND operator), otherwise a filter is carried out.

external

logical(1)
whether or not to return merely the table of external concepts.

match.input

logical(1)
whether or not to return a table that reflects the first input parameter exactly (default), or provides the unique items.

ontology

ontology(1)
either a path where the ontology is stored, or an already loaded ontology.

Value

A table of a subset of the ontology according to the values in ...

Examples

ontoDir <- system.file("extdata", "crops.rds", package = "ontologics")
onto <- load_ontology(path = ontoDir)

# exact matches from a loaded ontology ...
get_concept(label = "FODDER CROPS", ontology = onto)
#> # A tibble: 1 × 9
#>   id    label    description class has_broader has_broader_match has_close_match
#>   <chr> <chr>    <chr>       <chr> <chr>       <chr>             <chr>          
#> 1 .05   FODDER … NA          group NA          NA                NA             
#> # ℹ 2 more variables: has_exact_match <chr>, has_narrower_match <chr>

# ... or a path
get_concept(label = c("FODDER CROPS", "CEREALS"), ontology = ontoDir)
#> # A tibble: 2 × 9
#>   id    label    description class has_broader has_broader_match has_close_match
#>   <chr> <chr>    <chr>       <chr> <chr>       <chr>             <chr>          
#> 1 .05   FODDER … NA          group NA          NA                NA             
#> 2 .02   CEREALS  NA          group NA          NA                NA             
#> # ℹ 2 more variables: has_exact_match <chr>, has_narrower_match <chr>

# ignore querries that would not be valid in filter()
get_concept("label != 'Bioenergy woody' & has_broader == '.01'", "other", ontology = onto)
#> # A tibble: 2 × 9
#>   id     label   description class has_broader has_broader_match has_close_match
#>   <chr>  <chr>   <chr>       <chr> <chr>       <chr>             <chr>          
#> 1 .01.01 Bioene… NA          class .01         NA                NA             
#> 2 .01.03 Other … NA          class .01         NA                NA             
#> # ℹ 2 more variables: has_exact_match <chr>, has_narrower_match <chr>

# extract concepts based on regular expressions
get_concept(label = "crop", id = ".03$", regex = TRUE, ontology = ontoDir)
#> # A tibble: 1 × 9
#>   id    label description class has_broader has_broader_match has_close_match
#>   <chr> <chr> <chr>       <chr> <chr>       <chr>             <chr>          
#> 1 .03$  crop  NA          NA    NA          NA                NA             
#> # ℹ 2 more variables: has_exact_match <chr>, has_narrower_match <chr>