Get a concept in an ontology

Get a concept in an ontology

get_concept(
  terms = NULL,
  ...,
  regex = FALSE,
  tree = FALSE,
  missing = FALSE,
  ontology = NULL
)

Arguments

terms

character(1)

...

combination of column name and value to filter that column by. The value to filter by can be provided as regular expression.

regex

logical(1)
whether or not the value in ... shall be matched in full, or whether any partial match should be returned.

tree

logical(1)
whether or not to output the whole ontology tree starting from the given search terms.

missing

logical(1)
whether or not to give only those values that are currently missing from the ontology.

ontology

ontology(1)
either a path where the ontology is stored, or an already loaded ontology.

Value

A table of a subset of the ontology according to the values in ...

Examples

ontoDir <- system.file("extdata", "crops.rds", package = "ontologics")
onto <- load_ontology(path = ontoDir)

# exact matches from a loaded ontology ...
get_concept(terms = "FODDER CROPS", ontology = onto)
#> # A tibble: 1 × 6
#>   code  broader label_en     class external     sourceName
#>   <chr> <chr>   <chr>        <chr> <chr>        <chr>     
#> 1 .05   NA      FODDER CROPS group FODDER CROPS harmonised

# ... or one stored on the harddisc
get_concept(terms = "FODDER CROPS", ontology = ontoDir)
#> # A tibble: 1 × 6
#>   code  broader label_en     class external     sourceName
#>   <chr> <chr>   <chr>        <chr> <chr>        <chr>     
#> 1 .05   NA      FODDER CROPS group FODDER CROPS harmonised

# use regular expressions ...
get_concept(terms = "/*crops", regex = TRUE, ontology = onto)
#> # A tibble: 12 × 6
#>    code   broader label_en              class external sourceName
#>    <chr>  <chr>   <chr>                 <chr> <chr>    <chr>     
#>  1 .01.03 .01     Other bioenergy crops class NA       harmonised
#>  2 .03.01 .03     Fibre crops           class NA       harmonised
#>  3 .04.02 .04     Woody flower crops    class NA       harmonised
#>  4 .05.01 .05     Grass crops           class NA       harmonised
#>  5 .05.03 .05     Other fodder crops    class NA       harmonised
#>  6 .08.01 .08     Leguminous crops      class NA       harmonised
#>  7 .09.01 .09     Medicinal crops       class NA       harmonised
#>  8 .11.01 .11     Oilseed crops         class NA       harmonised
#>  9 .12.01 .12     Other crops           class NA       harmonised
#> 10 .14.01 .14     Stimulant crops       class NA       harmonised
#> 11 .14.02 .14     Spice crops           class NA       harmonised
#> 12 .15.03 .15     Other sugar crops     class NA       harmonised
get_concept(terms = "/*crops", broader = "_05", regex = TRUE, ontology = onto)
#> Warning: Using `as.character()` on a quosure is deprecated as of rlang 0.3.0.
#> Please use `as_label()` or `as_name()` instead.
#> This warning is displayed once per session.
#> # A tibble: 0 × 6
#> # … with 6 variables: code <chr>, broader <chr>, label_en <chr>, class <chr>,
#> #   external <chr>, sourceName <chr>

# get all concepts that are nested into another concept
get_concept(terms = "FODDER CROPS", tree = TRUE, ontology = onto)
#> # A tibble: 4 × 6
#>   code   broader label_en           class external sourceName
#>   <chr>  <chr>   <chr>              <chr> <chr>    <chr>     
#> 1 .05    NA      FODDER CROPS       group NA       harmonised
#> 2 .05.01 .05     Grass crops        class NA       harmonised
#> 3 .05.02 .05     Fodder legumes     class NA       harmonised
#> 4 .05.03 .05     Other fodder crops class NA       harmonised