Add a new (harmonised) concept to an ontology

Add a new (harmonised) concept to an ontology

new_concept(
  new,
  broader,
  class = NULL,
  source,
  overwrite = FALSE,
  attributes = NULL,
  ontoDir = NULL
)

Arguments

new

character(.)
the english label(s) of new concepts that shall be included in the ontology.

broader

character(.)
the english label(s) of already harmonised concepts to which the new concept shall be semantically linked via a skos:broader relation, see Details.

class

character(.)
the class(es) of the new labels.

source

character(1)
any character uniquely identifying the source dataset of the new concept (for example Author+Year).

overwrite

logical(1) ?

attributes

tibble()
not yet implemented.

ontoDir

character(1)
the path where the ontology in which to search is stored. It can be omitted in case the option "onto_path" has been define (see getOption("onto_path")).

Details

bla